add : SearchBar update : Network, SearchBar, UserForm, Navigation
Navigation 부분 아이템들 코드 최적화
