+++ src/components/Nav/NetWork.js
@@ -1,7 +1,29 @@
+import { useEffect, useState } from "react";
 import "./Network.css";
 
+const UseNetwork = (onChange) => {
+  const [status, setStatus] = useState(navigator.onLine);
+  const handleChange = () => {
+    if (typeof onChange === "function") {
+      onChange(navigator.onLine);
+    }
+  };
+  useEffect(() => {
+    window.addEventListener("online", handleChange);
+    window.addEventListener("offline", handleChange);
+    return () => {
+      window.removeEventListener("online", handleChange);
+      window.removeEventListener("offline", handleChange);
+    };
+  }, []);
+  return status;
+};
 const Network = () => {
-  return <div className="checknetwork" />;
+  const handleNetworkChange = (online) => {
+    console.log(online ? "온라인" : "오프라인");
+  };
+  const onLine = UseNetwork(handleNetworkChange);
+  return <>{onLine ? <div className="checknetwork" /> : null}</>;
 };
 
 export default Network;
