+++ src/page/Nav/Navigation.scss
@@ -25,18 +25,6 @@
   height: 82px;
 }
 
-.navigation-item-searchbar {
-  width: 700px;
-  height: 45px;
-  margin-left: 362px;
-  position: absolute;
-  outline: none;
-  border-top-left-radius: 50px;
-  border-bottom-left-radius: 50px;
-  border: 1px solid black;
-  font-size: 16px;
-}
-
 .navigation-item-search-off {
   transition: 0.5s;
   width: 60px;
